+++ draft-packages/stories/Collapsible.stories.tsx
@@ -1,6 +1,6 @@
 import { Icon, Paragraph } from "@kaizen/component-library"
 import { Collapsible, CollapsibleGroup } from "@kaizen/draft-collapsible"
-import { action } from "@storybook/addon-actions"
+
 import * as React from "react"
 import styles from "./Collapsible.stories.scss"
 import translationIcon from "@kaizen/component-library/icons/translation.icon.svg"
@@ -20,18 +20,6 @@ arcu quam a sapien. Donec in viverra urna.`
 
 export default {
   title: "Collapsible (React)",
-}
-
-export const SingleCollapsibleKaizenSiteDemo = () => (
-  <div style={{ margin: "1rem", width: "40rem" }}>
-    <Collapsible id="collapsible-single" open title="Single collapsible">
-      <Paragraph variant="body">{lipsum}</Paragraph>
-    </Collapsible>
-  </div>
-)
-
-SingleCollapsibleKaizenSiteDemo.story = {
-  name: "Single collapsible (Kaizen Site Demo)",
   component: Collapsible,
   parameters: {
     info: {
@@ -41,6 +29,16 @@ SingleCollapsibleKaizenSiteDemo.story = {
     },
   },
 }
+
+export const SingleCollapsibleKaizenSiteDemo = () => (
+  <div style={{ margin: "1rem", width: "40rem" }}>
+    <Collapsible id="collapsible-single" open title="Single collapsible">
+      <Paragraph variant="body">{lipsum}</Paragraph>
+    </Collapsible>
+  </div>
+)
+SingleCollapsibleKaizenSiteDemo.storyName =
+  "Single collapsible (Kaizen Site Demo)"
 
 export const SingleCollapsibleNoPadding = () => (
   <div style={{ margin: "1rem", width: "40rem" }}>
@@ -67,9 +65,7 @@ export const SingleCollapsibleNoPadding = () => (
   </div>
 )
 
-SingleCollapsibleNoPadding.story = {
-  name: "Single collapsible (no padding)",
-}
+SingleCollapsibleNoPadding.storyName = "Single collapsible (no padding)"
 
 export const SingleCollapsibleCustomHeader = () => (
   <div style={{ margin: "1rem", width: "40rem" }}>
@@ -95,9 +91,7 @@ export const SingleCollapsibleCustomHeader = () => (
   </div>
 )
 
-SingleCollapsibleCustomHeader.story = {
-  name: "Single collapsible (custom header)",
-}
+SingleCollapsibleCustomHeader.storyName = "Single collapsible (custom header)"
 
 export const SingleCollapsibleLazyLoad = () => (
   <div style={{ margin: "1rem", width: "40rem" }}>
@@ -115,9 +109,7 @@ export const SingleCollapsibleLazyLoad = () => (
   </div>
 )
 
-SingleCollapsibleLazyLoad.story = {
-  name: "Single collapsible (lazy load)",
-}
+SingleCollapsibleLazyLoad.storyName = "Single collapsible (lazy load)"
 
 // eslint-disable-next-line no-underscore-dangle
 export const _CollapsibleGroup = () => (
@@ -136,9 +128,7 @@ export const _CollapsibleGroup = () => (
   </div>
 )
 
-_CollapsibleGroup.story = {
-  name: "Collapsible group",
-}
+_CollapsibleGroup.storyName = "Collapsible group"
 
 export const CollapsibleGroupSeparated = () => (
   <div style={{ margin: "1rem", width: "40rem" }}>
@@ -156,9 +146,7 @@ export const CollapsibleGroupSeparated = () => (
   </div>
 )
 
-CollapsibleGroupSeparated.story = {
-  name: "Collapsible group (separated)",
-}
+CollapsibleGroupSeparated.storyName = "Collapsible group (separated)"
 
 export const CollapsibleGroupStickyHeaders = () => (
   <div style={{ margin: "4rem", width: "40rem" }}>
@@ -191,13 +179,11 @@ export const CollapsibleGroupStickyHeaders = () => (
   </div>
 )
 
-CollapsibleGroupStickyHeaders.story = {
-  name: "Collapsible group (sticky headers)",
-}
+CollapsibleGroupStickyHeaders.storyName = "Collapsible group (sticky headers)"
 
 export const CollapsibleGroupCallbackOnOpenClose = () => (
   <div style={{ margin: "1rem", width: "40rem" }}>
-    <CollapsibleGroup onToggle={action("Collapsible toggled")}>
+    <CollapsibleGroup onToggle={() => undefined}>
       <Collapsible id="collapsible-separate-1" open title="First panel">
         <Paragraph variant="body">{lipsum}</Paragraph>
       </Collapsible>
@@ -211,6 +197,5 @@ export const CollapsibleGroupCallbackOnOpenClose = () => (
   </div>
 )
 
-CollapsibleGroupCallbackOnOpenClose.story = {
-  name: "Collapsible group (callback on open/close)",
-}
+CollapsibleGroupCallbackOnOpenClose.storyName =
+  "Collapsible group (callback on open/close)"