+++ draft-packages/radio/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.30](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-radio@2.2.29...@kaizen/draft-radio@2.2.30) (2020-10-21)
+
+**Note:** Version bump only for package @kaizen/draft-radio
+
+
+
+
+
 ## [2.2.29](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-radio@2.2.28...@kaizen/draft-radio@2.2.29) (2020-10-08)
 
 **Note:** Version bump only for package @kaizen/draft-radio