+++ draft-packages/title-block-zen/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block-zen@2.3.1...@kaizen/draft-title-block-zen@2.4.0) (2020-10-21)
+
+
+### Features
+
+* Allow TitleBlockZen to collapse when no navigation ([#810](https://github.com/cultureamp/kaizen-design-system/issues/810)) ([442f952](https://github.com/cultureamp/kaizen-design-system/commit/442f952b73e492b57fc7a2e4df4a93c24f176bfb))
+
+
+
+
+
+## [2.3.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block-zen@2.3.0...@kaizen/draft-title-block-zen@2.3.1) (2020-10-21)
+
+**Note:** Version bump only for package @kaizen/draft-title-block-zen
+
+
+
+
+
+# [2.3.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block-zen@2.2.4...@kaizen/draft-title-block-zen@2.3.0) (2020-10-20)
+
+
+### Features
+
+* TitleBlockZen. Add a  variant for the badge ([#826](https://github.com/cultureamp/kaizen-design-system/issues/826)) ([e61961b](https://github.com/cultureamp/kaizen-design-system/commit/e61961b546ff0bcc5bd16c914d6147cf9fecca4c))
+
+
+
+
+
+## [2.2.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block-zen@2.2.3...@kaizen/draft-title-block-zen@2.2.4) (2020-10-19)
+
+
+### Bug Fixes
+
+* Move Title Block subcomponents out of peer deps ([#819](https://github.com/cultureamp/kaizen-design-system/issues/819)) ([4abec2f](https://github.com/cultureamp/kaizen-design-system/commit/4abec2f339fddabd17a64f9c06a81156cb67ac4f))
+
+
+
+
+
 ## [2.2.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block-zen@2.2.2...@kaizen/draft-title-block-zen@2.2.3) (2020-10-08)
 
 **Note:** Version bump only for package @kaizen/draft-title-block-zen