+++ packages/component-library/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.32.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@7.32.2...@kaizen/component-library@7.32.3) (2020-10-21)
+
+**Note:** Version bump only for package @kaizen/component-library
+
+
+
+
+
 ## [7.32.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@7.32.1...@kaizen/component-library@7.32.2) (2020-10-08)